utils/claude_client.py
from __future__ import annotations

import os
import time

import anthropic
from dotenv import load_dotenv

load_dotenv()

DEFAULT_MODEL = "claude-sonnet-4-6"
FAST_MODEL = "claude-haiku-4-5-20251001"

_client: anthropic.Anthropic | None = None


def _get_client() -> anthropic.Anthropic:
    global _client
    if _client is None:
        _client = anthropic.Anthropic(api_key=os.environ["ANTHROPIC_API_KEY"])
    return _client


def ask(
    prompt: str,
    max_tokens: int = 1024,
    model: str = DEFAULT_MODEL,
    system: str = "",
    retries: int = 3,
) -> str:
    """Call Claude and return the text response. Retries on transient errors."""
    messages = [{"role": "user", "content": prompt}]
    kwargs: dict = {"model": model, "max_tokens": max_tokens, "messages": messages}
    if system:
        kwargs["system"] = system

    for attempt in range(retries):
        try:
            resp = _get_client().messages.create(**kwargs)
            return resp.content[0].text
        except anthropic.RateLimitError:
            time.sleep(2 ** attempt * 5)
        except anthropic.APIStatusError as e:
            if e.status_code >= 500 and attempt < retries - 1:
                time.sleep(2 ** attempt * 2)
            else:
                raise
    raise RuntimeError(f"Claude call failed after {retries} retries")


def ask_fast(prompt: str, max_tokens: int = 512) -> str:
    """Use the fast/cheap model for low-stakes tasks."""
    return ask(prompt, max_tokens=max_tokens, model=FAST_MODEL)


def ask_with_tools(
    messages: list[dict],
    tools: list[dict],
    system: str = "",
    model: str = DEFAULT_MODEL,
    max_tokens: int = 4096,
    retries: int = 3,
) -> anthropic.types.Message:
    """
    Call Claude with tool definitions. Returns the full Message object.
    Used by subagents that need to inspect tool_use blocks in the response.
    """
    kwargs: dict = {
        "model": model,
        "max_tokens": max_tokens,
        "tools": tools,
        "messages": messages,
    }
    if system:
        kwargs["system"] = system

    for attempt in range(retries):
        try:
            return _get_client().messages.create(**kwargs)
        except anthropic.RateLimitError:
            time.sleep(2 ** attempt * 5)
        except anthropic.APIStatusError as e:
            if e.status_code >= 500 and attempt < retries - 1:
                time.sleep(2 ** attempt * 2)
            else:
                raise
    raise RuntimeError(f"Claude call failed after {retries} retries")
