requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
mcp>=1.0.0
uiautomator2>=3.2.0
pillow>=10.0.0
pixelmatch>=0.3.0
lxml>=5.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
aiohttp>=3.9.0
pyyaml>=6.0.0
